rag-project/requirements.txt
python-docx
PyMuPDF
sentence-transformers
faiss-cpu
ollama
langchain
torch
faiss-gpu